requirements.txt
matplotlib==3.8.2
scikit-learn==1.3.2
tensorflow==2.15.0
loguru==0.7.2
pyyaml==6.0.1
